CompData: fix inconsistent-result issue ISSUE=6218
When we have results, CompData returns a hash of results. However, when
we had no results (but didn't have any errors either), CompData returned
undef rather than an empty hash.

This commit changes things so that CompData returns an empty hash instead of
undef in this case, to make things simpler for consumers of CompData results.